+++ sdk/python/lib/test/test_policy.py
@@ -37,7 +37,8 @@ def run_policy(policy: Union[ResourceValidationPolicy, StackValidationPolicy]) -
     result = policy.validate(None, report)
     if isawaitable(result):
         loop = asyncio.new_event_loop()
-        loop.run_until_complete(result)
+        task = asyncio.Task(result, loop=loop) if asyncio.iscoroutine(result) else result
+        loop.run_until_complete(task)
         loop.close()
 
     return violations
@@ -46,8 +47,10 @@ def run_remediation(policy: ResourceValidationPolicy) -> Union[Mapping[str, Any]
     result = policy.remediate(None)
     if isawaitable(result):
         loop = asyncio.new_event_loop()
-        result = loop.run_until_complete(result)
+        task = asyncio.Task(result, loop=loop)
+        val = loop.run_until_complete(task)
         loop.close()
+        return val
 
     return result